+++ src/main/java/com/Service/dto/TopicDto.java
@@ -11,9 +11,9 @@ import java.util.Date;
 @ApiModel(value = "发布话题表数据传输对象")
 public class TopicDto implements Serializable {
     @ApiModelProperty(value = "话题id，主键自增",example = "1",position = 1,required = true)
-    private Integer id;
+    private Long id;
     @ApiModelProperty(value = "课程id，外键",example = "1",position = 2,required = true)
-    private Integer courseId;
+    private Long courseId;
     @ApiModelProperty(value = "创建者",example = "1",position = 3)
     private String createBy;
     @ApiModelProperty(value = "话题状态",example = "1",position = 4)
